+++ Cargo.toml
@@ -20,7 +20,7 @@ json-patch = "^4.0"
 log = "^0.4"
 rayon = "^1.10"
 regex = "^1.10.5"
-reqwest = "=0.12.15"
+reqwest = { version = "=0.12.15", features = ["blocking"] }
 semver = { version = "^1.0.23", features = ["serde"] }
 serde = { version = "^1.0.199", features = ["derive", "rc"] }
 serde_json = "^1.0.116"
