+++ src/server.rs
@@ -1,0 +1,4 @@
+mod api;
+mod index;
+
+pub use api::{ApiServer, ProcessingStatus};
